Continued implementation of Persistant Log In
Data is now persisted through the log in process. Usernames are required from new users before being able to see any additional content. The loading screen is now able to determine if a sub id is saved on the storage and/or if there is an existing account on the server with that sub id and what to do with all the inbetweens. Leo and I have spent a good amount of time testing this process for both persistance of data and edge use cases for potential errors/bugs. We have several checkpoints that will need to be addressed to concrete this initial auth rollout.

Co-Authored-By: lflores12 <lflores12@users.noreply.github.com>